+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 
 - Added a selection preview when using the Select Path tool.
+- Handles Prefab encapsulation with Probuilder Meshes
 
 ### Bug Fixes